сбор данных/requirements.txt
apache-airflow==2.8.0
kafka-python==2.0.2
pandas==2.0.3
sqlalchemy==1.4.53
requests==2.31.0
psycopg2-binary==2.9.9

